test(viewer): extract shared jsdom stubs + add unstubAllGlobals cleanup (review)
- Move the duplicated scroll/observer stubs into test-utils/jsdom-stubs.ts
  (stubBrowserAPIs), used by both player and dashboard tests.
- Stub per-test (beforeEach) and vi.unstubAllGlobals() in afterEach so stubbed
  globals don't leak between tests.

Refs #355
